test fixes

+++ packages/tddy-daemon/src/spawner.rs
@@ -354,7 +354,7 @@ pub fn clone_as_user(os_user: &str, git_url: &str, destination: &Path) -> anyhow
                 if libc::setgid(gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
-                if libc::initgroups(pw_name.as_ptr(), gid as libc::c_int) != 0 {
+                if libc::initgroups(pw_name.as_ptr(), gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
                 if libc::setuid(uid) != 0 {
@@ -581,7 +581,7 @@ pub fn spawn_as_user(
                 if libc::setgid(gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
-                if libc::initgroups(pw_name.as_ptr(), gid as libc::c_int) != 0 {
+                if libc::initgroups(pw_name.as_ptr(), gid) != 0 {
                     return Err(std::io::Error::last_os_error());
                 }
                 if libc::setuid(uid) != 0 {
